refactor: move report cards to package
