+++ unit/reportlist/src/commonMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/unit/reportlist/components/ReportCardPlaceHolder.kt
@@ -1,4 +1,4 @@
-package com.github.diegoberaldin.raccoonforlemmy.unit.reportlist
+package com.github.diegoberaldin.raccoonforlemmy.unit.reportlist.components
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
